使用vue-router，让程序支持新窗口渲染vue页面
