+++ package.json
@@ -43,6 +43,7 @@
     "jszip": "^3.10.1",
     "node-unrar-js": "^2.0.2",
     "pinia": "^2.1.7",
-    "regedit": "^5.1.3"
+    "regedit": "^5.1.3",
+    "vue-router": "4"
   }
 }
