+++ src/main.ts
@@ -5,6 +5,7 @@ import 'element-plus/dist/index.css'
 import '@imengyu/vue3-context-menu/lib/vue3-context-menu.css'
 import ContextMenu from '@imengyu/vue3-context-menu'
 import App from './App.vue'
+import { router } from './router'
 
 import './style.css'
 
@@ -13,6 +14,7 @@ import './demos/ipc'
 // import './demos/node'
 
 createApp(App)
+  .use(router)
   .use(createPinia())
   .use(ElementPlus)
   .use(ContextMenu)
