+++ src/windows/about.vue
@@ -1,0 +1,9 @@
+<script>
+</script>
+
+<template>
+    <div>这是about页面</div>
+</template>
+
+<style>
+</style>
